Create README.md
I'm currently studying GCSEs in maths and science so that I can be eligible for higher level courses. I'm also about to start a Higher Education Diploma with undergrad modules. After that I hope to take an MSc in Computer Science and Artificial Intelligence and then a PhD in Japan. That's why I'm taking relevant courses on Coursera.